Introduction to Crypto Bitcoin Trading with Python and Binance/requirements.txt
pandas
numpy
python-binance